refactor: colocate notification tracking hook under userTracking
